Add greater resilience to exception deserialisation errors.
If there is an exception during deserialisation of an exception,
previously this would silently kill the entire result receiving
thread.

This catches such exceptions, and passes that new exception along
as the result exception.

This also adds logging used to debug that.

The worker thread probably should have a logging wrapper round
the outside to catch any terminating exception and log it, so
at least it is clear the thread is dying; and also ahve the
same kind of exception around deserialisation of successful
results.