Always handle results via super() for more consistent / hopefully bug free experience
There's a callback race here demoed in test_pipeline.py

The handle_update callback is called, and a task is marked as completed, and so
we scan the list of tasks for tasks that have been unlocked.

However! We haven't, at that point, hit the AppFuture update callback that will
update the app future into being in done state (based on the parent callback).

Previously we have a consistency problem, maybe, where result() will return a
result even though the AppFuture is not marked as done - and we're relying on
that (racey?) behaviour to make progress.

I wonder if this can be trigger by the placement of a suitable sleep instead?
or if it is just reliant on the specific order in which callbacks fire?

This should be fine if we move to callback based launching of tasks, as the
launch will fire when the dependency is ready. But is there an incremental
way to get there?

Maybe we should also do a handle update when the app future is completed?
(although be careful of concurrency there?)

hande_update is fired when? when a *task* is completed

but data-base dependency tracking is waiting for what? when an *app* is completed.

Those are different things - and app completion is *after* task completion.


Breaks:   parsl/tests/test_bash_apps/test_pipeline.py

The above test hangs because of the above described race condition.