Some fiddling with datafuture done()
With the commit in place that changes the order in which app futures report their
done-ness wrt other callbacks, I now see this:

parsl.dataflow.dflow:215 [DEBUG] BENC: _count_deps checking dependency <DataFuture at 
0x7feaf27ca668 state=finished returned test1.txt> .done() == False

The state suggests the future is finished, but the result of done() does not.

That seems inconsistent.